Create two new exceptions, InvalidAttributeException and InvalidResourceException
No more magic resource and attribute handling, everything has to be specified now.
That makes a lot of the specs break i found out because of copy pasting stuff.
Also realised the normalize_hash method didn't return the "weird" uppercase attributes, so
{ "AmountVAT" => "12.2" } would return { amount_vat: "12.2" } instead of { amount_VAT: "12.2" }
This is a breaking change but for the better.